+++ packages/Panel/stories/examples/Showcase.js
@@ -15,7 +15,12 @@ const sidePanelProps = () => ({
   isCompact: boolean("isCompact", false, sidePanelGroup),
   isInline: boolean("isInline", false, sidePanelGroup),
   isOpen: boolean("isOpen", true, sidePanelGroup),
-  isSlideFromLeft: boolean("isSlideFromLeft", false, sidePanelGroup),
+  slideFrom: select(
+    "slideFrom",
+    [Panel.slideFromDirections.RIGHT, Panel.slideFromDirections.LEFT],
+    Panel.slideFromDirections.RIGHT,
+    sidePanelGroup
+  ),
   width: text("width", "50%", sidePanelGroup),
   zIndex: number("zIndex", undefined, {}, sidePanelGroup),
   offsetY: number("offsetY", 0, {}, sidePanelGroup),
@@ -38,7 +43,7 @@ const overlayProps = () => ({
   toggleOverlay: boolean("Include <Panel.Overlay>", true, overlayGroup),
 });
 
-const Showcase = props => {
+const Showcase = (props) => {
   return (
     <ExampleStory storyName="Panel" tagline={ExampleStory.defaultTaglines.showcase}>
       <Panel {...props}>